Refactor GripperCommand position to be joint instead of gripper gap size.
The GripperCommand position is in gap sizes because it was base on the PR2 gripper with
prismatic joints. The joints were proportinal with the gap size. For grippers with rotative
joints, setting a gap size dosen't represents the gripper visual state. The position is now
the desired joint position (rad).
See discussion: https://groups.google.com/forum/#!topic/moveit-users/qazNO7tmqOs